+++ src/Validator/Request/CookieValidator.php
@@ -135,7 +135,7 @@ final readonly class CookieValidator extends AbstractParameterValidator
             }
 
             $value = $this->deserializer->deserialize($value, $param);
-            $value = $this->coercer->coerce($value, $param, $this->coercion, true);
+            $value = $this->coercer->coerce($value, $param, $this->coercion, $this->config->strictCoercion);
 
             if (null !== $param->schema) {
                 $this->schemaValidator->validate($value, $param->schema, $context);